Create Skeleton for the action
This commit creates a test dummy skeleton that should be filled later on by the action.

The action currently gets the two parameters that are expected to be included and print them to the log.

Moreover, the README file is edited, and all related references to the boilerplate code are replaced