Add coverage badge to README
This commit appends the auto-generated coverage badge to the README file, so it is clear how much of the code is actually covered by tests.

This should give a sense of trust to the users that this action acutally tests the business logic behind the action, and for the contributors that dependencies could be updated automatically be dependabot without breaking the action.